更换数据库和日志

+++ app/common/type.go
@@ -15,7 +15,7 @@ func (d Date) MarshalJSON() ([]byte, error) {
 
 // BaseModel soft_delete 此插件还支持混合模式，删除同时更新删除时间
 type BaseModel struct {
-	ID        uint                  `gorm:"primaryKey" json:"id"`
+	ID        int64                 `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time             `json:"createdAt"`
 	UpdatedAt time.Time             `json:"-"`
 	IsDel     soft_delete.DeletedAt `gorm:"softDelete:flag;not null;default:0" json:"-"`
